requirements.txt
flask==3.0.3
werkzeug==3.0.3
numpy==1.26.4
pillow==10.3.0
rasterio==1.3.9
geopandas==0.15.2
shapely==2.0.4
scikit-learn==1.5.1
pyproj==3.6.1